dao: add dao for metadata tags
Includes a couple interesting patterns, namely index scanning with a
glob matcher, and index scanning with a prefix. Both of these can
probably be abstracted out into methods in @nothingness/level at some
point, and are pretty easy to write. Also introduced the use of a fast
hash to generate natural keys. One disadvantae of this approach is that
it's not very easy to scan by a prefix of the path, but this could
probably be added by combining prefix and glob-based searching, if it
turns out to be necessary.