dao: add a File DAO
Also, start dealing with the specific issues of de/serialization for
weird types in objects -- in this case dates. Serialization helpers
should probably be pulled up into the DAO base class, but only with the
knowledge that any mapping to constructors is handled in the
implementation of specific DAOs, and not by the DAO class.