dao: inject adaptor instead of initializing it
Because the DAOs need to call each other to reconstitute object graphs,
and because currently the whole cache lives in a single file (lol
performance), attempts to call between DAOs will fail because both have
access to the LevelDB instance simultaneously, which isn't allowed. So
just pass the adaptor to the file around instead.